feat(registrar): introduce temporary RegistrationBackend seam (CHAIN-4832)
Extract the Boundless register_signer workflow behind a cutover abstraction so hinted registration can land beside Boundless without stacking on planner/hints PRs. Production stays Boundless-only; the seam is removed with Boundless in CHAIN-4833.
